fixed bug where wouldn't play correct portion of the audio when clicking on a segment and wouldn't show the right red bar audio cursor location -- fix was to make sure we use 48K sample mp3 files
Former-commit-id: 829ea7d62664b15bbf88c03489216ade451130f8

+++ src/mitll/langtest/client/sound/SoundManagerMock.java
@@ -21,7 +21,7 @@ public class SoundManagerMock implements SoundManagerAPI {
     GWT.log("SoundManagerMock createSound");
   }
 
-  public void playInterval(Sound sound, double start, double end) {
+  public void playInterval(Sound sound, int start, int end) {
     //To change body of implemented methods use File | Settings | File Templates.
   }
